requirement.txt
Faker
snowflake-connector-python